% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{almost.lh.splits}
\alias{almost.lh.splits}
\title{Point clouds of peri-LH  neurons from split GAL4 lines}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 11.}
\usage{
almost.lh.splits
}
\description{
Point cloud objects from multi-colour flip-out image data. Meta-data contains experimentally verified information on neurostransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2017, upcoming.
}
\keyword{datasets}
