% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{capitalise_cell_type_name}
\alias{capitalise_cell_type_name}
\title{Capitalise cell type names}
\usage{
capitalise_cell_type_name(x, inverse = FALSE)
}
\arguments{
\item{x}{Cell type names}

\item{inverse}{If TRUE, makes the first two letters lowercase}
}
\description{
Capitalise the first two letters of the cell type names
}
