% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_summary}
\alias{cell_type_summary}
\title{Summary of cell types with clean split-GAL4 lines from Dolan et al. 2018}
\format{An object of class \code{data.frame} with 79 rows and 7 columns.}
\usage{
cell_type_summary
}
\description{
Collated by Dolan et al. 2018.
}
\keyword{datasets}
