% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{download_mophologies}
\alias{download_mophologies}
\title{Download all the LH neuronal skeletons in this package}
\usage{
download_mophologies(dir = paste0(getwd(), "/"), ...)
}
\arguments{
\item{dir}{path to directory into which to download the LH library}

\item{...}{additional arguments passed to nat::write.neuron}
}
\description{
Download all the neurons in the LH library as SWC files, compressed into a .zip
}
