% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{downloadskeletons}
\alias{downloadskeletons}
\title{Download neuronal skeletons and associated meta-data}
\usage{
downloadskeletons(nl, dir, format = "swc", subdir = NULL,
  INDICES = names(nl), files = NULL, Force = FALSE, ...)
}
\arguments{
\item{nl}{the neuronlist to download}

\item{dir}{directory to which to download data}

\item{format}{unique abbreviation of one of the registered file formats for neurons including 'swc', 'hxlineset', 'hxskel'}

\item{subdir}{string naming field in neuron that specifies a subdirectory OR expression to evaluate in the context of neuronlist's df attribute}

\item{INDICES}{character vector or expression specifying output filenames. See examples and nat::write.neuron for details}

\item{files}{character vector or expression specifying output filenames. See examples and nat::write.neuron for details}

\item{Force}{whether to overwrite an existing file}

\item{...}{additional arguments passed to nat::write.neuron}
}
\description{
Download neuronal morphologies as SWC files, compressed into a .zip file along with a CSV for their meta-data
}
