% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emlhns}
\alias{emlhns}
\title{EM skeletons for LHNs}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 143.}
\usage{
emlhns
}
\description{
EM skeletons for LHNs identified in the Dolan split lines. Tracing in progress.
}
\details{
Dolan et al. 2019 and Huoviala et al. 2019, upcoming.
}
\keyword{datasets}
