% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemibrain.lhn.bodyids}
\alias{hemibrain.lhn.bodyids}
\title{Body IDs for lateral horn neurons in the hemibrain volume.}
\format{
An object of class \code{numeric} of length 2027.
}
\source{
\url{https://doi.org/10.1101/2020.01.21.911859}
}
\usage{
hemibrain.lhn.bodyids
}
\description{
Body Ids for lateral horn neurons, pulled from neuPrint from the hemibrai volume. Only neurons marked by the project as 'Traced' have been considered.
Lateral horn neurons defined as third-order olfactory neurons with >= 1% of their postsynapses or >= 10 postsynapses supplied by canonical
olfacory uniglomerular PNs.
}
\references{
Shan Xu, C., Michal Januszewski, Zhiyuan Lu, Shin-Ya Takemura, Kenneth J. Hayworth, Gary Huang, Kazunori Shinomiya, et al. 2020.
“A Connectome of the Adult Drosophila Central Brain.”
bioRxiv. \href{https://doi.org/10.1101/2020.01.21.911859}{doi:10.1101/2020.01.21.911859}
}
\keyword{datasets}
