% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemibrain}
\alias{hemibrain}
\alias{hemibrain.lhn.bodyids}
\alias{hemibrain.ton.bodyids}
\alias{hemibrain_lhns}
\alias{hemibrain_pnt_cbf}
\alias{hemibrain_tons}
\title{Body IDs and FAFB matches for lateral horn neurons in the hemibrain volume.}
\format{
An object of class \code{numeric} of length 1635.

An object of class \code{character} of length 2361.

An object of class \code{data.frame} with 1634 rows and 23 columns.

An object of class \code{data.frame} with 76 rows and 5 columns.

An object of class \code{data.frame} with 76 rows and 5 columns.

An object of class \code{data.frame} with 4009 rows and 21 columns.
}
\source{
\url{https://doi.org/10.1101/2020.01.21.911859}
}
\usage{
hemibrain.lhn.bodyids

hemibrain.ton.bodyids

hemibrain_lhns

hemibrain_pnt_cbf

hemibrain_pnt_cbf

hemibrain_tons
}
\description{
Body Ids for lateral horn neurons, pulled from neuPrint from the hemibrain volume. Only neurons marked by the project as 'Traced' have been considered.
Lateral horn neurons defined as third-order olfactory neurons with >= 1% of their postsynapses or >= 10 postsynapses supplied by canonical
olfactory uniglomerular PNs. FAFB matches have been semi-manually found for some of these neurons.
}
\references{
Shan Xu, C., Michal Januszewski, Zhiyuan Lu, Shin-Ya Takemura, Kenneth J. Hayworth, Gary Huang, Kazunori Shinomiya, et al. 2020.
“A Connectome of the Adult Drosophila Central Brain.”
bioRxiv. \href{https://doi.org/10.1101/2020.01.21.911859}{doi:10.1101/2020.01.21.911859}
}
\keyword{datasets}
