% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jfw.lhns.dps}
\alias{jfw.lhns.dps}
\title{Vector clouds for lateral horn neuron morphologies from Jeanne and Fisek et al. 2018}
\format{An object of class \code{neuronlistfh} (inherits from \code{neuronlist}, \code{logical}) of length 377.}
\source{
\url{https://doi.org/10.1016/j.neuron.2018.05.011}
}
\usage{
jfw.lhns.dps
}
\description{
Skeletonized 3-D Morphologies of all Digitized LHNs related to Figure 4 in Jeanne and Fisek et al. 2018.
Meta-data contains the LHN's voltage responses to GH146 glomerular photostimulation. Transformed into FCWB space.
}
\seealso{
\code{\link{jfw.lhns}}
}
\keyword{datasets}
