% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh.em}
\alias{lh.em}
\alias{lh.fafb}
\alias{lh.hemibrain}
\alias{pn.fafb}
\alias{fafb_lhns}
\alias{lm_em_matches}
\title{EM neuron skeletons for published lateral horn related reconstructions and all olfactory projection neurons.}
\format{
An object of class \code{neuronlist} (inherits from \code{list}) of length 218.

An object of class \code{neuronlist} (inherits from \code{list}) of length 153.

An object of class \code{neuronlist} (inherits from \code{list}) of length 346.

An object of class \code{data.frame} with 218 rows and 20 columns.

An object of class \code{data.frame} with 2877 rows and 8 columns.
}
\source{
\url{https://doi.org/10.1101/2020.01.19.911453}
}
\usage{
lh.fafb

lh.hemibrain

pn.fafb

fafb_lhns

lm_em_matches
}
\description{
EM skeletons complete with synapses, a \code{neuronlist} objects for LH-associated neurons including projection neurons that input the lateral horn. These neurons have been reconstructed in both the FAFB ssTEM brain space
manually/semi-manually. They have also been cross-matched with their equivalent neurons from the FlyEM hemibrain project. These objects have been warped into the
\code{nat.flybrains::nat.flybrains::FCWB} template brain space, as has much of the data in this package. Note that the reconstruction of olfactory PN dendrites has not been done to synaptic-completion.
}
\details{
These neurons have also been split into their putative primary neurite, primary dendrite, dendrite, axon and soma compartments. See the function \code{hemibrainr::flow_centrality}
for details.
}
\references{
Bates, A. S., P. Schlegel, R. J. V. Roberts, N. Drummond, I. F.
  M. Tamimi, R. Turnbull, X. Zhao, et al. 2020. “Complete Connectomic
  Reconstruction of Olfactory Projection Neurons in the Fly Brain.” bioRxiv.
  \href{https://doi.org/10.1101/2020.01.19.911453}{doi:10.1101/2020.01.19.911453}.

Marin, Elizabeth C., Ruairí J. V. Roberts, Laurin Büld, Maria Theiss, Markus W. Pleijzier, Tatevik Sarkissian, Willem J. Laursen, et al. 2020.
“Connectomics Analysis Reveals First, Second, and Third Order Thermosensory and Hygrosensory Neurons in the Adult Drosophila Brain.”
bioRxiv.\href{https://doi.org/10.1101/2020.01.20.912709}{doi:10.1101/2020.01.20.912709}

Dolan, Michael-John, Shahar Frechter, Alexander Shakeel Bates, Chuntao Dan, Paavo Huoviala, Ruairí Jv Roberts, Philipp Schlegel, et al. 2019.
“Neurogenetic Dissection of the Drosophila Lateral Horn Reveals Major Outputs, Diverse Behavioural Functions, and Interactions with the Mushroom Body.” eLife 8 (May).
\href{https://doi.org/10.7554/eLife.43079}{doi:10.7554/eLife.43079}

Dolan, Michael-John, Ghislain Belliart-Guérin, Alexander Shakeel Bates, Shahar Frechter, Aurélie Lampin-Saint-Amaux, Yoshinori Aso, Ruairí J. V. Roberts, et al. 2018.
“Communication from Learned to Innate Olfactory Processing Centers Is Required for Memory Retrieval in Drosophila.”
Neuron 100 (3): 651–68.e8.\href{https://doi.org/10.1016/j.neuron.2018.08.037}{doi:10.1016/j.neuron.2018.08.037}

Huoviala, Paavo, Michael-John Dolan, Fiona M. Love, Shahar Frechter, Ruairí J. V. Roberts, Zane Mitrevica, Philipp Schlegel, et al. 2018.
“Neural Circuit Basis of Aversive Odour Processing in Drosophila from Sensory Input to Descending Output.”
bioRxiv. \href{https://doi.org/10.1101/394403}{doi:10.1101/394403}

Shan Xu, C., Michal Januszewski, Zhiyuan Lu, Shin-Ya Takemura, Kenneth J. Hayworth, Gary Huang, Kazunori Shinomiya, et al. 2020.
“A Connectome of the Adult Drosophila Central Brain.”
bioRxiv. \href{https://doi.org/10.1101/2020.01.21.911859}{doi:10.1101/2020.01.21.911859}

Li, Peter H., Larry F. Lindsey, Michał Januszewski, Zhihao Zheng, Alexander Shakeel Bates, István Taisz, Mike Tyka, et al. 2019.
“Automated Reconstruction of a Serial-Section EM Drosophila Brain with Flood-Filling Networks and Local Realignment.”
bioRxiv. \href{https://doi.org/10.1101/605634}{doi:10.1101/605634}
}
\keyword{datasets}
