% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh.inputs}
\alias{lh.inputs}
\title{A categorised database of neurons providing input to the lateral horn}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 1188.}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
lh.inputs
}
\description{
A categorised database of neurons providing input to the lateral horn
}
\details{
See \url{http://virtualflybrain.org} for definitions of tract
  abbreviations.
}
\examples{
\donttest{
library(nat)
head(lh.inputs)
# summarise numbers of neurons per modality/axon tract
with(lh.inputs, table(modality, tract))
table(most.lhns[,'pnt'])
}
}
\seealso{
\code{\link{most.lhns}}
}
\keyword{datasets}
