% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh.mcfo}
\alias{lh.mcfo}
\title{Single skeletons of LHNs from MCFO}
\format{An object of class \code{neuronlistfh} (inherits from \code{neuronlist}, \code{logical}) of length 538.}
\usage{
lh.mcfo
}
\description{
Single skeletons of LH neurons from split GAL4 line multi-colour-flip-out dat. Meta-data contains experimentally verified information on neurostransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2018, upcoming.
}
\keyword{datasets}
