% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh.splits.dps}
\alias{lh.splits.dps}
\title{Point clouds of LH output neurons from split GAL4 lines}
\format{An object of class \code{data.frame} with 220 rows and 20 columns.}
\usage{
lh.splits.dps
}
\description{
Point cloud objects from multi-colour flip-out image data. Meta-data contains experimentally verified information on neurostransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2017, upcoming.
}
\keyword{datasets}
