% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh_line_info}
\alias{lh_line_info}
\title{Split-Gal4 line codes for LHNs}
\format{An object of class \code{data.frame} with 404 rows and 19 columns.}
\usage{
lh_line_info
}
\description{
Collated by Dolan et al. 2018.
}
\keyword{datasets}
