% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh_tract_data}
\alias{lh_tract_data}
\title{Information about number/type of LHNs for each tract}
\format{An object of class \code{data.frame} with 21 rows and 5 columns.}
\usage{
lh_tract_data
}
\description{
Information about number/type of LHNs for each tract
}
\details{
This was based on work by Arian Jamasb, Ruari Roberts, Alex Bates,
  and Greg Jefferis as a contribution to Frechter et al. 2018.

  The sampling strategy counted profiles in all the main LH associated tracts
  and then used random sampling and tracing to the first branchpoint to
  estimate the number of LHNs in each tract.

  From LHN_Tract_Table_Tex.R @ b49e87462c938222313e4f6835572bbdb2fb5727
}
\seealso{
\code{\link{primary.neurite.tracts}}
}
\keyword{datasets}
