% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhin.splits.dps}
\alias{lhin.splits.dps}
\title{Vector clouds of LH input neurons from split GAL4 lines}
\format{
An object of class \code{neuronlist} (inherits from \code{list}) of length 42.

An object of class \code{neuronlist} (inherits from \code{list}) of length 42.
}
\source{
\url{https://doi.org/10.7554/eLife.43079}

\url{https://doi.org/10.7554/eLife.43079}
}
\usage{
lhin.splits.dps

lhin.splits.dps
}
\description{
Vector cloud objects from segmented image data. Meta-data contains experimentally verified information on neurotransmitter expression. Data from Mike Dolan.

Vector cloud objects from segmented image data. Meta-data contains experimentally verified information on neurotransmitter expression. Data from Mike Dolan.
}
\references{
Dolan, Michael-John, Shahar Frechter, Alexander Shakeel Bates, Chuntao Dan, Paavo Huoviala, Ruairí Jv Roberts, Philipp Schlegel, et al. 2019.
“Neurogenetic Dissection of the Drosophila Lateral Horn Reveals Major Outputs, Diverse Behavioural Functions, and Interactions with the Mushroom Body.” eLife 8 (May).
\href{https://doi.org/10.7554/eLife.43079}{doi:10.7554/eLife.43079}

Dolan, Michael-John, Shahar Frechter, Alexander Shakeel Bates, Chuntao Dan, Paavo Huoviala, Ruairí Jv Roberts, Philipp Schlegel, et al. 2019.
“Neurogenetic Dissection of the Drosophila Lateral Horn Reveals Major Outputs, Diverse Behavioural Functions, and Interactions with the Mushroom Body.” eLife 8 (May).
\href{https://doi.org/10.7554/eLife.43079}{doi:10.7554/eLife.43079}
}
\keyword{datasets}
