% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhln.splits.dps}
\alias{lhln.splits.dps}
<<<<<<< HEAD
\title{Vector clouds of LH local neurons from split GAL4 lines}
\format{An object of class \code{neuronlistfh} (inherits from \code{neuronlist}, \code{logical}) of length 60.}
=======
\title{Point clouds of LH local neurons from split GAL4 lines}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 57.}
>>>>>>> origin/master
\usage{
lhln.splits.dps
}
\description{
<<<<<<< HEAD
Vector cloud objects from segmented image data. Meta-data contains experimentally verified information on neurostransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2018, upcoming.
=======
Point cloud objects from multi-colour flip-out image data. Meta-data contains experimentally verified information on neurostransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2017, upcoming.
>>>>>>> origin/master
}
\keyword{datasets}
