% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhns-package.r
\docType{package}
\name{lhns}
\alias{lhns}
\title{Skeleton data for Drosophila melanogaster lateral horn (associated) neurons}
\description{
The key data objects provided by this package is the \code{\link{most.lhns}}
containing neurons
}
