% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{light.pn.axons}
\alias{light.pn.axons}
\title{Light level tracign of PN axons}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 242.}
\usage{
light.pn.axons
}
\description{
Light level tracings of PN axons in FCWB space from Jefferis and Potter et al. 2007, Wong and Wang et al. 2002 and Yu et al. 2010.
}
\keyword{datasets}
