% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbons.light}
\alias{mbons.light}
\title{Mushroom body output neuron vector clouds representing}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 17.}
\usage{
mbons.light
}
\description{
Point cloud objects from segmented image data. In FCWB space. Data from Yoshi Aso.
}
\details{
Aso et al. 2014 \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4273437/}
}
\keyword{datasets}
