% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{most.lhins.dps}
\alias{most.lhins.dps}
\title{A categorised database of Vector clouds for neurons providing input to the lateral horn}
\format{An object of class \code{neuronlistfh} (inherits from \code{neuronlist}, \code{logical}) of length 1197.}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
most.lhins.dps
}
\description{
A categorised database of Vector clouds for neurons providing input to the lateral horn
}
\details{
See \url{http://virtualflybrain.org} for definitions of tract
  abbreviations.
}
\examples{
\donttest{
library(nat)
head(most.lhins)
# summarise numbers of neurons per modality/axon tract
with(most.lhins, table(modality, tract))
table(most.lhns[,'pnt'])
}
}
\seealso{
\code{\link{most.lhns}}
}
\keyword{datasets}
