% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{most.lhins}
\alias{most.lhins}
\title{A categorised database of neurons providing input to the lateral horn}
\format{
An object of class \code{neuronlist} (inherits from \code{list}) of length 1258.
}
\source{
\url{http://www.flycircuit.tw/}

\url{http://www.flycircuit.tw/}
}
\usage{
most.lhins
}
\description{
A categorised database of neurons providing input to the lateral horn
}
\details{
See \url{http://virtualflybrain.org} for definitions of tract
  abbreviations.
}
\examples{
\donttest{
library(nat)
head(most.lhins)
# summarise numbers of neurons per modality/axon tract
with(most.lhins, table(modality, tract))
table(most.lhns[,'pnt'])
}

}
\references{
Frechter, Shahar, Alexander Shakeel Bates, Sina Tootoonian, Michael-John Dolan, James D. Manton, Arian Rokkum Jamasb, Johannes Kohl, Davi Bock, and Gregory Sxe Jefferis. 2019.
“Functional and Anatomical Specificity in a Higher Olfactory Centre.” eLife 8 (May).
\href{https://doi.org/10.7554/eLife.44590}{doi:10.7554/eLife.44590}
}
\seealso{
\code{\link{most.lhns}}
}
\keyword{datasets}
