% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pn.axons.light}
\alias{pn.axons.light}
\title{Light level tracing of PN axons}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 222.}
\usage{
pn.axons.light
}
\description{
Light level tracings of PN axons in FCWB space from Jefferis and Potter et al. 2007, Wong and Wang et al. 2002 and Yu et al. 2010.
}
\keyword{datasets}
