% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_lhn_name}
\alias{process_lhn_name}
\title{Get primary neurite tract name and anatomy group from cell types}
\usage{
process_lhn_name(x)
}
\arguments{
\item{x}{Cell type names}
}
\description{
Get primary neurite tract name and anatomy group from cell types
}
