% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{see_fc}
\alias{see_fc}
\title{See a FlyCircuit neuron on the FlyCircuit website}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
see_fc(fc_id, max = 10)
}
\arguments{
\item{fc_id}{A FlyCircuit neuron ID, either the neuron ID or the gene name ID}

\item{max}{Maximum number of webpages to open}
}
\description{
See chosen FlyCircuit neuron on the FlyCircuit website, where maximal projection .lsm images are available.
}
