% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{badel.PN.responses}
\alias{badel.PN.responses}
\title{Projection neuron response data}
\format{An object of class \code{matrix} with 84 rows and 37 columns.}
\source{
\url{http://dx.doi.org/10.1016/j.neuron.2016.05.022}
}
\usage{
badel.PN.responses
}
\description{
Ca2+ responses in the dendrites of Drosophila, uniglomerular GH146 positive
projection neurons from Badel et al. 2017.
}
\keyword{datasets}
