% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emlhns.dps}
\alias{emlhns.dps}
\title{Vector clouds of EM skeletons for LHNs}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 78.}
\usage{
emlhns.dps
}
\description{
Vector clouds for EM LHNs identified in the Dolan split lines. Tracing in progress.
}
\details{
Dolan et al. 2018, upcoming.
}
\keyword{datasets}
