% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh.mcfo.dps}
\alias{lh.mcfo.dps}
\title{Vector clouds for single LHNs from MCFO}
\format{An object of class \code{neuronlistfh} (inherits from \code{neuronlist}, \code{logical}) of length 377.}
\usage{
lh.mcfo.dps
}
\description{
Single skeletons of LH neurons from split GAL4 line multi-colour-flip-out dat. Meta-data contains experimentally verified information on neurostransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2018, upcoming.
}
\keyword{datasets}
