% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh.splits.dps}
\alias{lh.splits.dps}
\title{Vector clouds of LH output neurons from split GAL4 lines}
\format{
An object of class \code{neuronlist} (inherits from \code{list}) of length 220.
}
\source{
\url{https://doi.org/10.7554/eLife.43079}
}
\usage{
lh.splits.dps
}
\description{
Vector cloud objects from segmented image data. Meta-data contains experimentally verified information on neurotransmitter expression. As of v2.0 of this package, the names from these neurons have been
modified from those used in Frechter et al. 2019 due to changes in the LH naming scheme as it was
exapnded to all lateral horn neurons within the hemibrain (Scheffer et al. 2020). Data from Mike Dolan.
}
\references{
Dolan, Michael-John, Shahar Frechter, Alexander Shakeel Bates, Chuntao Dan, Paavo Huoviala, Ruairí Jv Roberts, Philipp Schlegel, et al. 2019.
“Neurogenetic Dissection of the Drosophila Lateral Horn Reveals Major Outputs, Diverse Behavioural Functions, and Interactions with the Mushroom Body.” eLife 8 (May).
\href{https://doi.org/10.7554/eLife.43079}{doi:10.7554/eLife.43079}
}
\keyword{datasets}
