% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh_line_info}
\alias{lh_line_info}
\title{Gal4 line codes for LHNs}
\format{An object of class \code{data.frame} with 510 rows and 16 columns.}
\usage{
lh_line_info
}
\description{
Collated in Frechter et al. 2018.
}
\keyword{datasets}
