% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lh_nblast}
\alias{lh_nblast}
\title{Pre-calculated NBLAST scores between all neurons in this package}
\format{An object of class \code{matrix} with 2142 rows and 2142 columns.}
\usage{
lh_nblast
}
\description{
Saves time re-claculating an NBLAST beetween all LH associated neuron morphologies
}
\keyword{datasets}
