% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhin.splits.dps}
\alias{lhin.splits.dps}
\title{Vector clouds of LH input neurons from split GAL4 lines}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 42.}
\usage{
lhin.splits.dps

lhin.splits.dps
}
\description{
Vector cloud objects from segmented image data. Meta-data contains experimentally verified information on neurotransmitter expression. Data from Mike Dolan.

Vector cloud objects from segmented image data. Meta-data contains experimentally verified information on neurotransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2018, upcoming.

Dolan et al. 2018, upcoming.
}
\keyword{datasets}
