% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhn_odour_responses}
\alias{lhn_odour_responses}
\title{Odour responses for different LHN cell types}
\format{An object of class \code{matrix} with 143 rows and 36 columns.}
\usage{
lhn_odour_responses
}
\description{
LHNs can be found in most.lhns. From Frechter et al. 2018
}
\keyword{datasets}
