% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhns.pn.overlap.matrix}
\alias{lhns.pn.overlap.matrix}
\title{Dendritic overlap of LHN cell types with axon terminals of PN inputs to LH}
\format{An object of class \code{matrix} with 1494 rows and 270 columns.}
\usage{
lhns.pn.overlap.matrix
}
\description{
This matrix is organised by the identifier of the PN inputs (rows) and the
LHN cell type (columns).
}
\seealso{
\code{\link{most.lhns}}, \code{\link{most.lhins}}
}
\keyword{datasets}
