% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhon.splits.dps}
\alias{lhon.splits.dps}
\title{Vector clouds of LH output neurons from split GAL4 lines}
\format{An object of class \code{neuronlistfh} (inherits from \code{neuronlist}, \code{logical}) of length 136.}
\usage{
lhon.splits.dps
}
\description{
Vector cloud objects from segmented image data. Meta-data contains experimentally verified information on neurostransmitter expression. Data from Mike Dolan.
}
\details{
Dolan et al. 2018, upcoming.
}
\keyword{datasets}
