% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{most.lhins}
\alias{most.lhins}
\title{A categorised database of neurons providing input to the lateral horn}
\format{An object of class \code{data.frame} with 1188 rows and 13 columns.}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
most.lhins
}
\description{
A categorised database of neurons providing input to the lateral horn
}
\details{
See \url{http://virtualflybrain.org} for definitions of tract
  abbreviations.
}
\examples{
\donttest{
library(nat)
head(most.lhins)
# summarise numbers of neurons per modality/axon tract
with(most.lhins, table(modality, tract))
table(most.lhns[,'pnt'])
}
}
\seealso{
\code{\link{most.lhns}}
}
\keyword{datasets}
