% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{most.lhns.dps}
\alias{most.lhns.dps}
\title{A comprehensive lateral horn neuron dotprops dataset}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 1172.}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
most.lhns.dps
}
\description{
A \code{\link[nat]{dotprops}} dataset containing all of the FlyCircuit
lateral horn neurons that seem to have dendrite in the LH, as well as traced
dye-fills from Frechter et al. 2017
}
\seealso{
\code{\link{most.lhns}}
}
\keyword{datasets}
