% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{most.lhns}
\alias{most.lhns}
\title{A comprehensive lateral horn neuron dataset}
\format{
An object of class \code{neuronlist} (inherits from \code{list}) of length 1646.
}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
most.lhns
}
\description{
A dataset containing all of the FlyCircuit neurons that seem to have
dendrites in the lateral horn (LH), as well as traced dye-fills from Frechter
et al. 2019. The neurons are in the \code{nat} package's
\code{\link[nat]{neuron}} format. As of v2.0 of this package, the names from these neurons have been
modified from those used in Frechter et al. 2019 due to changes in the LH naming scheme as it was
exapnded to all lateral horn neurons within the hemibrain (Scheffer et al. 2020).
}
\examples{
\donttest{
library(nat)
head(most.lhns)
# summarise numbers of neurons per Primary Neurite tract
table(most.lhns[,'pnt'])
}
\dontrun{
good_lhns=subset(most.lhns, good.trace & !is.na(anatomy.group) &
                 anatomy.group!="notLHproper")
write.neurons(good_lhns, "/path/to/ouput/dir", subdir=file.path(pnt,anatomy.group),
              files=paste0(id, '.swc'))
}
}
\references{
Frechter, Shahar, Alexander Shakeel Bates, Sina Tootoonian, Michael-John Dolan, James D. Manton, Arian Rokkum Jamasb, Johannes Kohl, Davi Bock, and Gregory Sxe Jefferis. 2019.
“Functional and Anatomical Specificity in a Higher Olfactory Centre.” eLife 8 (May).
\href{https://doi.org/10.7554/eLife.44590}{doi:10.7554/eLife.44590}
}
\seealso{
\code{\link{most.lhns.dps}}
}
\keyword{datasets}
