% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{most.lhns}
\alias{most.lhns}
\title{A comprehensive lateral horn neuron dataset}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 1226.}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
most.lhns
}
\description{
A dataset containing all of the FlyCircuit neurons that seem to have
dendrites in the lateral horn (LH), as well as traced dye-fills from Frechter
et al. 2017. The neurons are in the \code{nat} package's
\code{\link[nat]{neuron}} format.
}
\examples{
\donttest{
library(nat)
head(most.lhns)
# summarise numbers of neurons per Primary Neurite tract
table(most.lhns[,'pnt'])
}
\dontrun{
good_lhns=subset(most.lhns, good.trace & !is.na(anatomy.group) &
                 anatomy.group!="notLHproper")
write.neurons(good_lhns, "/path/to/ouput/dir", subdir=file.path(pnt,anatomy.group),
              files=paste0(id, '.swc'))
}
}
\seealso{
most.lhns.dps
}
\keyword{datasets}
