% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/olf.info.R
\docType{data}
\name{olf.info}
\alias{olf.info}
\alias{olf.info}
\title{Information on glomeruli and their cognate olfactory sensaroy neurons and receptors}
\format{An object of class \code{data.frame} with 70 rows and 3 columns.}
\usage{
olf.info

olf.info
}
\description{
Information on glomeruli and their cognate olfactory sensaroy neurons and receptors from DoOR.

Information on glomeruli and their cognate olfactory sensaroy neurons and receptors from DoOR.
}
\details{
Information collated in: \url{https://www.ncbi.nlm.nih.gov/pubmed/27653699}
}
\keyword{datasets}
