% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnts.R
\name{plot_pnts}
\alias{plot_pnts}
\title{Plot adult Lateral Horn primary neurite tracts}
\usage{
plot_pnts(open = FALSE, brain = nat.flybrains::FCWB)
}
\arguments{
\item{open}{whether to open a new RGL window for plotting}

\item{brain}{the brainspace in which to plot the primary neurite tracts. Defaults to FCWB
(default=\code{TRUE})}
}
\description{
Plots the primary neurite tracts from Frechter et al. 2017 in
  \code{\link[nat.flybrains]{FCWB}} space.
}
\seealso{
\code{\link[nat.flybrains]{FCWB}}
}
