% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pn.info}
\alias{pn.info}
\title{Information on projection neurons}
\format{An object of class \code{data.frame} with 77 rows and 14 columns.}
\usage{
pn.info
}
\description{
Information on projection neurons, their glomeruli and their cognate olfactory sensaroy neurons and receptors.
}
\details{
Data largely from \url{https://www.ncbi.nlm.nih.gov/pubmed/27653699}
}
\keyword{datasets}
