% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary.neurite.tracts}
\alias{primary.neurite.tracts}
\title{Lateral horn primary neurite tracts as a dotprops object}
\format{
An object of class \code{neuronlist} (inherits from \code{list}) of length 31.
}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
primary.neurite.tracts
}
\description{
The known primary neurite tracts that lateral horn neurons can take from
their somata, from Frechter et al. 2018.
}
\examples{
\donttest{
library(nat)
plot_pnts()
plot3d(most.lhns, pnt=='AV6', col='black', soma=TRUE)

## a bit messy, so maybe try just with the spine (longest path of neuron)
# removes neurons
npop3d()
# nb UseStartPoint => longest path from soma
av6.spines=nlapply(subset(most.lhns, pnt=='AV6'), spine, UseStartPoint=TRUE)
plot3d(av6.spines, col='black')
}

}
\references{
Frechter, Shahar, Alexander Shakeel Bates, Sina Tootoonian, Michael-John Dolan, James D. Manton, Arian Rokkum Jamasb, Johannes Kohl, Davi Bock, and Gregory Sxe Jefferis. 2019.
“Functional and Anatomical Specificity in a Higher Olfactory Centre.” eLife 8 (May).
\href{https://doi.org/10.7554/eLife.44590}{doi:10.7554/eLife.44590}
}
\seealso{
\code{\link{plot_pnts}}, \code{\link{lh_tract_data}}
}
\keyword{datasets}
