% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary.neurite.tracts}
\alias{primary.neurite.tracts}
\title{Lateral horn primary neurite tracts as a dotprops object}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 30.}
\usage{
primary.neurite.tracts
}
\description{
The known primary neurite tracts that lateral horn neurons can take from
their somata, from Frechter et al. 2018.
}
\examples{
\donttest{
library(nat)
plot_pnts()
plot3d(most.lhns, pnt=='av6', col='black', soma=T)

## a bit messy, so maybe try just with the spine (longest path of neuron)
# removes neurons
npop3d()
# nb UseStartPoint => longest path from soma
av6.spines=nlapply(subset(most.lhns, pnt=='av6'), spine, UseStartPoint=T)
plot3d(av6.spines, col='black')
}
}
\seealso{
\code{\link{plot_pnts}}, \code{\link{lh_tract_data}}
}
\keyword{datasets}
